pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "CDaniel"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-admin (>=2.0.2,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "neo4j (>=5.28.2,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
